core-api/src/requirements.txt
Flask==3.0.0
Flask-Cors==5.0.0
Flask-JWT-Extended==4.5.3
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
gunicorn==21.2.0
python-dotenv==1.0.0
marshmallow-sqlalchemy==0.29.0
requests==2.31.0
marshmallow==3.20.1
ollama==0.4.7
SQLAlchemy==2.0.21